+++ spec/recurly/coupon_spec.rb
@@ -28,4 +28,15 @@ describe Coupon do
       coupon.redeem 'xX_pinkman_Xx'
     end
   end
+
+  describe "#unlimited_redemptions_per_account?" do
+    it "should be false unless max_redemptions_per_account is nil" do
+      coupon.max_redemptions_per_account = 10
+      coupon.unlimited_redemptions_per_account?.must_equal false
+    end
+    it "should be true of max_redemptions_per_account is nil" do
+      coupon.max_redemptions_per_account = nil
+      coupon.unlimited_redemptions_per_account?.must_equal true
+    end
+  end
 end